requirements.txt
bip_utils==1.7.0
bitsv==0.11.5
cashaddress==1.0.6
aioeos==1.0.2
pywallet==0.1.0
monero==0.8
PyWaves==0.8.15
base58==2.0.0 #Override for EOS

opencv-python==4.5.1.48
arrow==0.17.0
asn1crypto==1.4.0
async_timeout==3.0.1
asyncgui==0.1.3
asynctkinter==0.1.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==6.6
coincurve==15.0.0
colorama==0.4.4
docker_py==1.8.0
ecdsa==0.15
flake8==3.8.4
funcsigs==1.0.2
idna==2.10
iniconfig==1.1.1
jsonschema==3.2.0
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.1
opencv_python==4.5.1.48
packaging==20.9
path.py==12.5.0
path==15.1.2
pexpect==4.8.0
Pillow==8.1.0
pip==21.0.1
pluggy==0.13.1
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
pyflakes==2.2.0
pygame==2.0.1
pyparsing==2.4.7
pytest==6.2.2
python_dateutil==2.8.1
PyYAML==5.4.1
pyzbar==0.1.8
requests==2.25.1
setuptools==53.0.0
six==1.15.0
tabulate==0.8.9
toml==0.10.2
typing_extensions==3.7.4.3
urllib3==1.26.3
websocket_client==0.57.0
yarl==1.6.3
future==0.18.2
imutils==0.5.4
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
mnemonic==0.13
pbkdf2==1.3
protobuf==3.0.0a3
pyaes==1.6.1
pyblake2==1.1.2
pyrsistent==0.17.3
pysha3==1.0.2
python_axolotl_curve25519==0.4.1.post2
python_bitcoinrpc==1.0
sha256==0.1
two1==3.10.9
varint==1.0.2
whatsonchain==0.0.3

mnemonic==0.19
cbor==1.0.0
scrypt==0.8.18
ed25519==1.5